vertaisarvioitava

+++ hdr/aiMove.h
@@ -10,6 +10,6 @@ void GenerateMovesWhite(uint32_t *moves, uint64_t pawns, uint64_t knights,
 void GenerateMovesBlack(uint32_t *moves, uint64_t pawns, uint64_t knights,
 						uint64_t bishops, uint64_t rooks, uint64_t queens,
 						uint64_t kings, uint64_t fBoard, uint64_t eBoard);
-uint16_t GetMove(uint64_t boards[14], bool white);
+uint32_t GetMove(uint64_t boards[14], bool white);
 
 #endif
